separate isTarget functions
